use component for asking email
